foodconnect-backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.22
SQLAlchemy>=2.0
pydantic>=2.4
pydantic-settings>=2.2
passlib[bcrypt]>=1.7
psycopg[binary]>=3.1
python-jose[cryptography]>=3.3
